refactor(orgs): Separate membership seats and option max_seats
