+++ ams/memberships/management/commands/create_sample_membership_options.py
@@ -52,7 +52,6 @@ class Command(management.base.BaseCommand):
                 "duration": relativedelta(months=1),
                 "cost": 99.99,
                 "invoice_reference": "AMS Membership",
-                "max_seats": 5,
                 "archived": False,
             },
         )
@@ -63,18 +62,17 @@ class Command(management.base.BaseCommand):
                 "duration": relativedelta(years=1),
                 "cost": 999.99,
                 "invoice_reference": "AMS Membership",
-                "max_seats": 10,
                 "archived": False,
             },
         )
         MembershipOption.objects.update_or_create(
-            name="Sample Organisation Membership - Unlimited",
+            name="Sample Organisation Membership - 1 Year (Max 5 seats)",
             type=MembershipOptionType.ORGANISATION,
             defaults={
                 "duration": relativedelta(years=1),
-                "cost": 2499.99,
+                "cost": 599.99,
                 "invoice_reference": "AMS Membership",
-                "max_seats": None,  # No limit
+                "max_seats": 5,
                 "archived": False,
             },
         )
