+++ ams/organisations/email_utils.py
@@ -173,7 +173,7 @@ def send_staff_organisation_membership_notification(membership):
         "membership": membership,
         "organisation": membership.organisation,
         "membership_option": membership.membership_option,
-        "seats": membership.max_seats,
+        "seats": membership.seats,
         "start_date": membership.start_date,
         "expiry_date": membership.expiry_date,
         "cost": membership.membership_option.cost,
@@ -253,7 +253,7 @@ def send_staff_organisation_seats_added_notification(
         "membership": membership,
         "membership_option": membership.membership_option,
         "seats_added": seats_added,
-        "new_total_seats": membership.max_seats,
+        "new_total_seats": membership.seats,
         "prorata_cost": prorata_cost,
         "invoice_number": invoice.invoice_number if invoice else None,
         "expiry_date": membership.expiry_date,
